requirements.txt
absl-py==0.10.0
actionlib==1.12.1
angles==1.9.12
attrs==19.3.0
bondpy==1.8.5
branca @ file:///home/conda/feedstock_root/build_artifacts/branca_1588860304135/work
brotlipy==0.7.0
cachetools==4.1.1
camera-calibration==1.15.0
camera-calibration-parsers==1.11.13
catkin==0.7.29
certifi==2020.6.20
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1595805769248/work
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
controller-manager==0.18.2
controller-manager-msgs==0.18.2
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1595348770643/work
cudnnenv==0.7.0
cv-bridge==1.13.0
cycler==0.10.0
decorator==4.4.2
descartes==1.1.0
diagnostic-analysis==1.9.7
diagnostic-common-diagnostics==1.9.7
diagnostic-updater==1.9.7
dynamic-reconfigure==1.6.3
efficientnet-pytorch==0.7.0
Fiona==1.8.13
folium @ file:///home/conda/feedstock_root/build_artifacts/folium_1588861378954/work
future==0.18.2
gazebo-plugins==2.8.7
gazebo-ros==2.8.7
GDAL==3.1.2
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.14
geopandas @ file:///home/conda/feedstock_root/build_artifacts/geopandas_1594925563431/work
google-auth==1.22.1
google-auth-oauthlib==0.4.1
grpcio==1.32.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
image-geometry==1.13.0
imageio==2.9.0
interactive-markers==1.11.5
Jinja2==2.11.2
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1593624380152/work
joint-state-publisher==1.12.15
kdl-parser-py==1.13.1
keyboard==0.13.5
kiwisolver==1.2.0
laser-geometry==1.6.5
Markdown==3.3.2
MarkupSafe==1.1.1
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-base_1597428381320/work
message-filters==1.14.10
MouseInfo==0.1.3
munch==2.5.0
networkx==2.4
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1595522867729/work
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.44
osmnx @ file:///home/conda/feedstock_root/build_artifacts/osmnx_1593803983597/work
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1595958450914/work
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1594213113401/work
pretrainedmodels==0.7.4
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.1
pyproj @ file:///home/conda/feedstock_root/build_artifacts/pyproj_1595910437254/work
PyQt5==5.15.1
PyQt5-sip==12.8.1
PyRect==0.1.4
PyScreeze==0.1.26
PySocks==1.7.1
python-dateutil==2.8.1
python-qt-binding==0.4.3
python3-xlib==0.15
PyTweening==1.0.3
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1592425495151/work
requests-oauthlib==1.3.0
resource-retriever==1.12.6
rosbag==1.14.10
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.10
roslaunch==1.14.10
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.10
rosmake==1.14.9
rosmaster==1.14.10
rosmsg==1.14.10
rosnode==1.14.10
rosparam==1.14.10
rospy==1.14.10
rosservice==1.14.10
rostest==1.14.10
rostopic==1.14.10
rosunit==1.14.9
roswtf==1.14.10
rqt-action==0.4.9
rqt-bag==0.5.0
rqt-bag-plugins==0.5.0
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.9
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.2
rqt-py-console==0.4.8
rqt-reconfigure==0.5.3
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.6.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rsa==4.6
Rtree==0.9.4
rviz==1.13.15
scikit-image==0.17.2
scikit-learn @ file:///home/conda/feedstock_root/build_artifacts/scikit-learn_1596546073704/work
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1595583577416/work
sensor-msgs==1.12.7
Shapely==1.7.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
threadpoolctl @ file:///tmp/tmp79xdzxkt/threadpoolctl-2.1.0-py3-none-any.whl
tifffile==2020.12.8
topic-tools==1.14.10
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tornado==6.0.4
tqdm==4.50.2
typing-extensions==3.7.4.3
urdfdom-py==0.4.4
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1595434816409/work
Werkzeug==1.0.1
xacro==1.13.9
